Fetching meals data optimization

+++ backend/src/main/java/com/jacek/nutriweek/menu/dto/MealItemDTO.java
@@ -1,4 +1,4 @@
 package com.jacek.nutriweek.menu.dto;
 
 public record MealItemDTO (ProductDTO product,
-                           float amount){}
+                           Double amount){}
